CXR_Classifier/requirements.txt
# download dataset
kagglehub==0.3.12

# Core dependencies
numpy==2.2.6
tqdm==4.67.1
psutil==7.0.0

# Machine Learning and Deep Learning
scikit-learn==1.7.0
torch==2.7.1
torchmetrics==1.7.3
torchvision==0.22.1

# Experiment tracking and configuration
hydra-core==1.3.2
mlflow==2.21.3
pip==25.1.1 # helps MLflow detect pip version

# Logging
python-json-logger==3.3.0

# Visualization
matplotlib==3.10.3
seaborn==0.13.2

# Image processing
Pillow==11.3.0

